requirements.txt
# ============================================================================
# MediaForge Studio - Dependencies
# ============================================================================
# Python 3.10+ required

# ----------------------------------------------------------------------------
# UI Framework
# ----------------------------------------------------------------------------
PySide6>=6.6.0
qfluentwidgets>=1.4.0
qdarktheme>=2.1.0

# ----------------------------------------------------------------------------
# Video/Audio Processing
# ----------------------------------------------------------------------------
yt-dlp>=2024.1.0
ffmpeg-python>=0.2.0
pydub>=0.25.1

# ----------------------------------------------------------------------------
# AI/ML - Speech Recognition
# ----------------------------------------------------------------------------
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numba==0.60.0  # Whisper compatibility fix

# ----------------------------------------------------------------------------
# TTS Engines
# ----------------------------------------------------------------------------
edge-tts>=6.1.9                    # Default TTS (no external engine required)
pyttsx3>=2.90                      # Offline TTS fallback
# google-cloud-texttospeech>=2.14.0  # Optional - Google Cloud TTS (requires API key)

# ----------------------------------------------------------------------------
# Translation APIs
# ----------------------------------------------------------------------------
aiohttp>=3.9.0                     # For API requests
requests>=2.31.0                   # For sync requests
deepl>=1.15.0                      # DeepL translation
argostranslate>=1.9.0              # Offline translation

# ----------------------------------------------------------------------------
# Configuration & Utilities
# ----------------------------------------------------------------------------
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2

# ----------------------------------------------------------------------------
# Data Handling & Progress
# ----------------------------------------------------------------------------
pydantic>=2.5.0
tqdm>=4.66.0
Pillow>=10.0.0

# ----------------------------------------------------------------------------
# Build & Distribution
# ----------------------------------------------------------------------------
pyinstaller>=6.0.0

# ============================================================================
# Development Dependencies (Optional)
# ============================================================================
# Uncomment these for development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.1.0
# mypy>=1.7.0

# ============================================================================
# Optional External Dependencies
# ============================================================================
# - FFmpeg: Required for video/audio processing (install separately)
#   Windows: choco install ffmpeg
#   macOS: brew install ffmpeg
#   Linux: sudo apt install ffmpeg
#
# - VOICEVOX Engine: Optional for high-quality Japanese TTS
#   Download from: https://voicevox.hiroshiba.jp/
#   Start VOICEVOX Engine before using this TTS option
#
# - API Keys (configure in .env file):
#   - GROQ_API_KEY: Free AI translation (recommended)
#   - GEMINI_API_KEY: Google Gemini translation
#   - DEEPL_API_KEY: DeepL translation
#   - OPENROUTER_API_KEY: OpenRouter AI translation
#   - OPENAI_API_KEY: OpenAI translation
# ============================================================================
